refactor(oauth): Divide component to gratify useSearchParams Suspense Boundary wrapping rule
