+++ apps/client/src/app/oauth/kakao/page.tsx
@@ -1,45 +1,11 @@
-'use client';
+import { Suspense } from 'react';
 
-import { useRouter, useSearchParams } from 'next/navigation';
-import { useEffect } from 'react';
-
-import { setTokenCookie, setUserInfoCookie } from '@/auth/cookie/setCookie';
-import OAuthRedirect from '@/components/features/oauth/OAuthRedirect';
-import { postKakaoAuthCode } from '@/services/api/oauth';
-import { useUserInfo } from '@/stores/userInfo';
+import KakaoRedirectClient from './KakaoRedirectClient';
 
 export default function KakaoRedirectPage() {
-    const searchParams = useSearchParams();
-    const router = useRouter();
-
-    useEffect(() => {
-        const handleKakaoLogin = async () => {
-            const code = searchParams.get('code');
-
-            if (code) {
-                const serverResponse = await postKakaoAuthCode(code);
-                if (!serverResponse) return;
-
-                const { data: loginUserInfo, accessToken } = serverResponse;
-
-                useUserInfo.getState().setUserInfo(loginUserInfo);
-
-                // 🍪 쿠키 관련 임시 설정 (추후 refactor 필요)
-                setTokenCookie(accessToken);
-                setUserInfoCookie();
-
-                router.push(
-                    loginUserInfo.enabled
-                        ? '/main'
-                        : loginUserInfo.nickName
-                          ? '/auth/terms'
-                          : '/auth/nickname'
-                );
-            }
-        };
-
-        handleKakaoLogin();
-    }, [searchParams, router]);
-
-    return <OAuthRedirect />;
+    return (
+        <Suspense fallback={<div>Loading...</div>}>
+            <KakaoRedirectClient />
+        </Suspense>
+    );
 }
